requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7

# AI/ML dependencies
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
scikit-learn==1.3.2

# Data processing
pandas==2.1.3

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Word cloud generation
wordcloud==1.9.2
Pillow==10.1.0

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
loguru==0.7.2
jmespath==1.0.1
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Optional: For better performance
# accelerate==0.24.1
# optimum==1.14.1
